feat: simple empty state
