+++ src/lib/pages/_components/dashboard/Banner.svelte
@@ -1,0 +1,17 @@
+<script lang="ts">
+	import { SuiteFavicons } from '@ramo-libre/ui-themes';
+</script>
+
+<div
+	class="bg-linear-to-br from-primary-100 to-primary-100/80 rounded-xl p-8 shadow-sm text-base-100"
+>
+	<div class="flex items-center gap-4">
+		<img src={SuiteFavicons.web} alt="Ramo Libre Logo" class="w-14 h-14" />
+		<div>
+			<h1 class="text-2xl lg:text-3xl font-bold tracking-tight">Ramo Libre</h1>
+			<p class="text-base-100/80 text-sm lg:text-base mt-1">
+				Gestiona tus ramos, horarios y notas de forma inteligente.
+			</p>
+		</div>
+	</div>
+</div>
