+++ src/lib/pages/_components/dashboard/OtherApps.svelte
@@ -1,0 +1,39 @@
+<script lang="ts">
+	import { ExternalLink } from '@lucide/svelte';
+	import { SuiteFavicons } from '@ramo-libre/ui-themes';
+
+	const apps = [
+		{
+			icon: SuiteFavicons.lab,
+			name: 'Ramo Libre LAB',
+			desc: 'Para probar combinaciones de notas.',
+			href: 'https://lab.ramolibre.app',
+			id: 'lab'
+		},
+		{
+			icon: SuiteFavicons.hub,
+			name: 'Ramo Libre HUB',
+			desc: 'Busca en tu comunidad informacion de ramos.',
+			href: 'https://hub.ramolibre.app',
+			id: 'hub'
+		}
+	];
+</script>
+
+<div class="grid grid-cols-2 gap-3">
+	{#each apps as app (app.id)}
+		<a
+			href={app.href}
+			target="_blank"
+			rel="noopener noreferrer"
+			class="group bg-base-100 border border-base-400 rounded-xl p-4 shadow-sm hover:bg-base-200 transition-colors text-left flex items-center gap-3"
+		>
+			<img src={app.icon} alt={app.name} class="w-8 h-8 shrink-0" />
+			<div class="min-w-0 flex-1">
+				<h3 class="text-sm lg:text-base font-bold text-content truncate">{app.name}</h3>
+				<p class="text-xs text-content/60 mt-0.5 line-clamp-2">{app.desc}</p>
+			</div>
+			<ExternalLink class="w-4 h-4 text-content/30 group-hover:text-primary-100 shrink-0 transition-colors" />
+		</a>
+	{/each}
+</div>
